backend/requirements.txt
flask
flask-cors
flask-executor
flask-sqlalchemy
pydantic
python-dotenv
openai
anthropic
reportlab
requests